+++ rate-service/src/main/java/com/arqui/rateservice/entity/Rate.java
@@ -16,6 +16,6 @@ public class Rate {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     private long id;
-    private Integer rate;
-    private Integer rate_extra_pause;
+    private double rate;
+    private double rate_extra_pause;
 }
